+++ projects/w11k/tydux/src/lib/Entity.store.test.ts
@@ -1,78 +1,69 @@
-describe("empty test", () => {
-    it("should do nothing", () => {
-        // success('Noop');
-    });
-});
-/*
-import {assert} from "chai";
-import {enableTyduxDevelopmentMode} from "./development";
-import {EntityStore} from "./Entity.store";
-import {resetTydux} from "./global-state";
-import {afterAllStoreEvents, collect} from "./test-utils";
+import {EntityStoreFacade} from "./Entity.store";
+import {createTyduxStore} from "./store";
 
-class MyEnt {
-    constructor(readonly id: number, public fieldA: string) {
+class Person {
+    constructor(readonly id: number, readonly name: string) {
     }
 }
 
-describe("EntityStore", function () {
+describe("EntityStoreFacade", () => {
 
-    beforeEach(() => enableTyduxDevelopmentMode());
-
-    afterEach(() => resetTydux());
-
-    it("instances can be added with `add`", function () {
-        let es = new EntityStore("es1", MyEnt, "id");
-
-        es.add(new MyEnt(1, "a"));
-        const entity = es.state.entities["1"];
-        assert.equal(entity.id, 1);
-        assert.equal(entity.fieldA, "a");
+    it("state fields are null initially", () => {
+        const tydux = createTyduxStore();
+        const esf = new EntityStoreFacade<Person>(tydux.createMountPoint("persons"), "id");
+        expect(esf.state.list).toBeNull();
+        expect(esf.state.byId).toBeNull();
     });
 
-    it("selectById()", async function () {
-        let es = new EntityStore("es1", MyEnt, "id");
-        es.add(new MyEnt(1, "a"));
+    it("isPristine() at start", () => {
+        const tydux = createTyduxStore();
+        const esf = new EntityStoreFacade<Person>(tydux.createMountPoint("persons"), "id");
+        expect(esf.isPristine).toBeTruthy();
+    });
 
-        await afterAllStoreEvents(es);
+    it("is not pristine once an entity was added", () => {
+        const tydux = createTyduxStore();
+        const esf = new EntityStoreFacade<Person>(tydux.createMountPoint("persons"), "id");
+        const p0 = new Person(0, "Joe");
+        esf.add(p0);
+        expect(esf.isPristine()).toBeFalsy();
+    });
 
-        let changes = collect(es.selectById("1"));
+    it("is pristine after reset()", () => {
+        const tydux = createTyduxStore();
+        const esf = new EntityStoreFacade<Person>(tydux.createMountPoint("persons"), "id");
+        const p0 = new Person(0, "Joe");
+        esf.add(p0);
+        expect(esf.isPristine()).toBeFalsy();
+        esf.reset();
+        expect(esf.isPristine()).toBeTruthy();
+    });
 
-        changes.assert(
-            new MyEnt(1, "a")
+    it("add entry with idSelector function", () => {
+        const tydux = createTyduxStore();
+        const esf = new EntityStoreFacade<Person>(
+            tydux.createMountPoint("persons"),
+            e => e.id
         );
+        const p0 = new Person(0, "Joe");
+        esf.add(p0);
+        expect(esf.state.list).toEqual([p0]);
+        expect(esf.state.byId).toEqual({0: p0});
     });
 
-    it("selectById() emits undefined until ID exists", async function () {
-        let es = new EntityStore("es1", MyEnt, "id");
-        let changes = collect(es.selectById("1"));
-        es.add(new MyEnt(1, "a"));
-
-        await afterAllStoreEvents(es);
-
-        changes.assert(
-            undefined,
-            new MyEnt(1, "a")
+    it("add entry with idSelector field", () => {
+        const tydux = createTyduxStore();
+        const esf = new EntityStoreFacade<Person>(
+            tydux.createMountPoint("persons"),
+            "id"
         );
+        const p0 = new Person(0, "Joe");
+        esf.add(p0);
+        expect(esf.state.list).toEqual([p0]);
+        expect(esf.state.byId).toEqual({0: p0});
     });
 
-    it("selectAll()", async function () {
-        let es = new EntityStore("es1", MyEnt, "id");
-        let changes = collect(es.selectAll());
 
-        es.add(new MyEnt(1, "a"));
-        es.add(new MyEnt(2, "b"));
-        es.add(new MyEnt(3, "c"));
 
-        await afterAllStoreEvents(es);
-
-        changes.assert(
-            [],
-            [new MyEnt(1, "a")],
-            [new MyEnt(1, "a"), new MyEnt(2, "b")],
-            [new MyEnt(1, "a"), new MyEnt(2, "b"), new MyEnt(3, "c")]
-        );
-    });
 
 });
-*/