+++ projects/w11k/tydux/src/lib/Facade.test.ts
@@ -1,10 +1,11 @@
-import {Action, createStore, Store as ReduxStore} from "redux";
+import {createStore, Store as ReduxStore} from "redux";
 import {distinctUntilChanged, map} from "rxjs/operators";
 import {createAsyncPromise, createTestMount} from "../testing";
 import {collect} from "../testing/test-utils-internal";
 import {Commands} from "./commands";
+import {enableTyduxDevelopmentMode} from "./development";
 import {Facade} from "./Facade";
-import {createTyduxStore, MountPoint, TyduxReducerBridge, TyduxStore} from "./store";
+import {createTyduxStore, TyduxReducerBridge, TyduxStore} from "./store";
 import {areArraysShallowEquals, untilNoBufferedStateChanges} from "./utils";
 
 
@@ -16,7 +17,8 @@ describe("Facade", () => {
         class TestFacade extends Facade<any, any> {
         }
 
-        const tf1 = new TestFacade(store.createMountPoint("testFacade"), null, Commands);
+        // tslint:disable-next-line:no-unused-expression
+        new TestFacade(store.createMountPoint("testFacade"), null, Commands);
         expect(() => {
             // tslint:disable-next-line
             new TestFacade(store.createMountPoint("testFacade"), null, Commands);
@@ -24,17 +26,16 @@ describe("Facade", () => {
     });
 
     it("can be destroyed", () => {
-        const mount = createTestMount({});
         let called = false;
 
         class TestFacade extends Facade<any, any> {
-            constructor(mountPoint: MountPoint<any, any>, name: string, commands: any) {
-                super(mountPoint, name, commands);
-                this.destroyed.subscribe(() => called = true);
+            constructor() {
+                super(createTestMount(), {}, Commands);
+                this.observeDestroyed().subscribe(() => called = true);
             }
         }
 
-        const facade = new TestFacade(mount, "TestFacade", Commands);
+        const facade = new TestFacade();
         facade.destroy();
         expect(called).toBe(true);
     });
@@ -52,13 +53,10 @@ describe("Facade", () => {
             }
         }
 
-        const mount = createTestMount({n1: 0});
-        const facade = new TestFacade(mount, "TestFacade", new TestCommands());
+        const facade = new TestFacade(createTestMount({n1: 0}), undefined, new TestCommands());
 
         const values: any = [];
-        facade.select((currentState) => {
-            values.push(currentState);
-        }).subscribe();
+        facade.select((currentState) => values.push(currentState)).subscribe();
 
         facade.actionInc();
         facade.actionInc();
@@ -85,7 +83,7 @@ describe("Facade", () => {
             }
         }
 
-        const facade = new TestFacade(createTestMount({n1: 0}), "TestFacade", new TestCommands());
+        const facade = new TestFacade(createTestMount({n1: 0}), undefined, new TestCommands());
         const collected = collect(facade.select(s => s.n1));
         facade.actionInc();
         facade.actionInc();
@@ -116,7 +114,7 @@ describe("Facade", () => {
             }
         }
 
-        const facade = new TestFacade(createTestMount({a: 0, b: 10, c: 100}), "TestFacade", new TestCommands());
+        const facade = new TestFacade(createTestMount({a: 0, b: 10, c: 100}), undefined, new TestCommands());
         const collected = collect(facade.select(s => [s.a, s.b]));
         facade.actionIncAB();
         facade.actionIncC();
@@ -149,7 +147,7 @@ describe("Facade", () => {
             }
         }
 
-        const facade = new TestFacade(createTestMount({a: 0, b: 10, c: 100}), "TestFacade", new TestCommands());
+        const facade = new TestFacade(createTestMount({a: 0, b: 10, c: 100}), undefined, new TestCommands());
         const collected = collect(facade.select(s => {
             return {
                 a: s.a,
@@ -182,7 +180,7 @@ describe("Facade", () => {
         }
 
         const state = {root: {child: {val1: 1}}};
-        const store = new TestFacade(createTestMount(state), "TestFacade", new TestCommands());
+        const store = new TestFacade(createTestMount(state), undefined, new TestCommands());
         const collected = collect(store.select(s => s.root));
         store.action(); // should not trigger select()
         store.action(); // should not trigger select()
@@ -216,7 +214,7 @@ describe("Facade", () => {
             }
         }
 
-        const facade = new TestFacade(createTestMount(new TestState()), "TestFacade", new TestCommands());
+        const facade = new TestFacade(createTestMount(new TestState()), undefined, new TestCommands());
         const collected = collect(facade.select(s => s.count));
         facade.action();
 
@@ -224,8 +222,22 @@ describe("Facade", () => {
         collected.assert(0, 1, 2, 1);
     });
 
+    it("select() first emits the current state", function (done) {
+        const initialState = {
+            val: 10
+        };
+
+        const tyduxStore = createTyduxStore(initialState);
+
+        tyduxStore.select().subscribe(state => {
+            expect(state.val).toEqual(10);
+            done();
+        });
+    });
+
     it("keeps state between action invocations", async () => {
         class TestState {
+            // noinspection JSMismatchedCollectionQueryUpdate
             list: number[] = [];
             value?: number;
         }
@@ -251,7 +263,7 @@ describe("Facade", () => {
         }
 
 
-        const facade = new TestFacade(createTestMount(new TestState()), "TestFacade", new TestCommands());
+        const facade = new TestFacade(createTestMount(new TestState()), undefined, new TestCommands());
         facade.setList();
         facade.setValue();
 
@@ -279,37 +291,16 @@ describe("Facade", () => {
             }
         }
 
-        const facade = new TestFacade(createTestMount(new TestState()), "TestFacade", new TestCommands());
+        const facade = new TestFacade(createTestMount(new TestState()), undefined, new TestCommands());
         facade.met1();
-    });
-
-    it("should provide an overloaded constructor to pass TyduxStore instead of mount point", () => {
-        class TestState {
-        }
-
-        class TestCommands extends Commands<TestState> {
-        }
-
-        class TestFacade extends Facade<TestState, TestCommands> {
-        }
-
-        const tyduxBridge = new TyduxReducerBridge();
-        const reduxStore = createStore(tyduxBridge.createTyduxReducer(), {});
-        const tydux = tyduxBridge.connectStore(reduxStore);
-
-        const facade = new TestFacade(tydux, "test", new TestCommands(), new TestState());
-
-        expect(reduxStore.getState()).toEqual({test: {}});
-
-        expect(facade.state).toEqual({});
     });
 
     it("can set their initial state during super call", () => {
         type TestState = { value: number };
 
         class TestFacade extends Facade<TestState, Commands<TestState>> {
-            constructor(tydux: TyduxStore<any>) {
-                super(tydux.createMountPoint("test"), "test", new Commands(), {value: 0});
+            constructor(tyduxStore: TyduxStore) {
+                super(tyduxStore.createMountPoint("test"), {value: 0}, new Commands());
 
             }
         }
@@ -331,13 +322,45 @@ describe("Facade", () => {
         });
     });
 
+    it("undefined as initialState does not alter the state", () => {
+        type TestState = { value: number };
+
+        class TestFacade extends Facade<TestState, Commands<TestState>> {
+            constructor(tyduxStore: TyduxStore) {
+                super(tyduxStore.createMountPoint("test"), undefined, new Commands());
+
+            }
+        }
+
+        const tyduxBridge = new TyduxReducerBridge();
+        const reduxStore = createStore(tyduxBridge.createTyduxReducer({
+            global: true,
+            test: {
+                value: 0
+            }
+        }));
+        const tydux = tyduxBridge.connectStore(reduxStore);
+        const facade = new TestFacade(tydux);
+
+        expect(reduxStore.getState()).toEqual({
+            global: true,
+            test: {
+                value: 0
+            }
+        } as any);
+
+        expect(facade.state).toEqual({
+            value: 0
+        });
+    });
+
     it("can set their initial state during super call with a function", () => {
 
         type TestState = { value: number };
 
         class TestFacade extends Facade<TestState, Commands<TestState>> {
-            constructor(tydux: TyduxStore<any>) {
-                super(tydux.createMountPoint("test"), "test", new Commands(), () => ({value: 99}));
+            constructor(tyduxStore: TyduxStore) {
+                super(tyduxStore.createMountPoint("test"), () => ({value: 99}), new Commands());
 
             }
         }
@@ -359,7 +382,7 @@ describe("Facade", () => {
         });
     });
 
-    it("can set their initial state during super call with a proise", (done) => {
+    it("can set their initial state during super call with a promise", (done) => {
 
         type TestState = { value: number };
 
@@ -368,8 +391,8 @@ describe("Facade", () => {
         });
 
         class TestFacade extends Facade<TestState, Commands<TestState>> {
-            constructor(tydux: TyduxStore<any>) {
-                super(tydux.createMountPoint("test"), "test", new Commands(), initialValuePromise);
+            constructor(tyduxStore: TyduxStore) {
+                super(tyduxStore.createMountPoint("test"), initialValuePromise, new Commands());
 
             }
         }
@@ -413,7 +436,7 @@ describe("Facade", () => {
             }
         }
 
-        const store = new TestFacade(createTestMount(new TestState()), "TestFacade", new TestCommands());
+        const store = new TestFacade(createTestMount(new TestState()), undefined, new TestCommands());
         try {
             store.op();
         } catch (e) {
@@ -452,7 +475,7 @@ describe("Facade", () => {
             }
         }
 
-        const store = new TestFacade(createTestMount(new TestState()), "TestFacade", new TestCommands());
+        const store = new TestFacade(createTestMount(new TestState()), undefined, new TestCommands());
         await store.setList();
         await store.setValue();
 
@@ -471,9 +494,9 @@ describe("Facade", () => {
         }
 
         const tyduxBridge = new TyduxReducerBridge();
-        const reduxStore: ReduxStore<typeof initialState, Action> = createStore(tyduxBridge.wrapReducer(plainReducer));
+        const reduxStore: ReduxStore = createStore(tyduxBridge.wrapReducer(plainReducer));
         const connected = tyduxBridge.connectStore(reduxStore);
-        const mount = connected.internalCreateMountPoint(s => s, (_, s) => ({...s}));
+        const mount = connected.createMountPoint("TestMount");
 
         class TestCommands extends Commands<{ list1: number[], list2: number[] }> {
             setList1(list: number[]) {
@@ -490,7 +513,7 @@ describe("Facade", () => {
             private counter = 0;
 
             constructor() {
-                super(mount, "", new TestCommands());
+                super(mount, initialState, new TestCommands());
 
                 this.select()
 
@@ -533,3 +556,173 @@ describe("Facade", () => {
     });
 
 });
+
+describe("Facade - sanity tests", function () {
+
+    beforeEach(() => enableTyduxDevelopmentMode());
+
+    it("can not modify the state directly", function () {
+        class TestFacade extends Facade<any, any> {
+            action() {
+                (this.state as any).count = 1;
+            }
+        }
+
+        const facade = new TestFacade(createTestMount(), undefined, Commands);
+        expect(() => facade.action()).toThrow();
+    });
+
+    it("can not assign the state", function () {
+        class TestFacade extends Facade<any, any> {
+            action() {
+                (this.state as any) = {};
+            }
+        }
+
+        const facade = new TestFacade(createTestMount(), undefined, Commands);
+        expect(() => facade.action()).toThrow();
+    });
+
+    it("member method can use async/await", async function (done) {
+        class MyState {
+            count = 0;
+        }
+
+        class TestCommands extends Commands<MyState> {
+            incrementBy(by: number) {
+                this.state.count += by;
+            }
+        }
+
+        class TestFacade extends Facade<MyState, TestCommands> {
+            async action() {
+                this.commands.incrementBy(1);
+                const by = await createAsyncPromise(10);
+                this.commands.incrementBy(by);
+
+                await untilNoBufferedStateChanges(facade);
+
+                collected.assert(0, 1, 11);
+                done();
+            }
+        }
+
+        const facade = new TestFacade(createTestMount(new MyState()), undefined, new TestCommands());
+        const collected = collect(facade.select(s => s.count));
+        facade.action();
+    });
+
+    it("member method can use member variables", function () {
+        class TestFacade extends Facade<any, any> {
+
+            counterA?: number;
+
+            counterB = 1;
+
+            action() {
+                this.counterA = 10;
+                this.innerAction();
+            }
+
+            private innerAction() {
+                this.counterB = 20;
+            }
+        }
+
+        const facade = new TestFacade(createTestMount(), undefined, Commands);
+        facade.action();
+        expect(facade.counterA).toEqual(10);
+        expect(facade.counterB).toEqual(20);
+    });
+
+    it("member method can use async/await and instance variables", function (done) {
+        class TestFacade extends Facade<any, any> {
+
+            counter = 0;
+
+            async action() {
+                this.counter = 10;
+                await createAsyncPromise(10);
+                this.counter++;
+
+                setTimeout(() => {
+                    expect(this.counter).toEqual(11);
+                    done();
+                }, 0);
+
+            }
+        }
+
+        const store = new TestFacade(createTestMount(), undefined, Commands);
+        store.action();
+    });
+
+    it("member methods and invoked sibling methods access the same instance variables", function () {
+        class TestFacade extends Facade<any, any> {
+
+            counter = 0;
+
+            action() {
+                this.counter = 10;
+                this.check();
+            }
+
+            private check() {
+                expect(this.counter).toEqual(10);
+            }
+        }
+
+        const store = new TestFacade(createTestMount(), undefined, Commands);
+        store.action();
+    });
+
+    it("member method can use async/await and call sibling methods", function (done) {
+        class TestFacade extends Facade<any, any> {
+
+            chars = "A";
+
+            async action() {
+                this.chars += "B";
+                this.append();
+                await createAsyncPromise(10);
+                this.chars += "C";
+                this.append();
+                this.chars += "E";
+
+                setTimeout(() => {
+                    expect(this.chars).toEqual("ABXCXE");
+                    done();
+                }, 0);
+            }
+
+            private append() {
+                this.chars += "X";
+            }
+        }
+
+        const facade = new TestFacade(createTestMount(), undefined, Commands);
+        facade.action();
+    });
+
+    it("exception in action method does not revert changes to instance variables", function () {
+        class TestFacade extends Facade<any, any> {
+
+            chars = "";
+
+            action() {
+                this.chars = "A";
+                throw new Error();
+            }
+        }
+
+        const facade = new TestFacade(createTestMount(), undefined, Commands);
+
+        try {
+            facade.action();
+        } catch (e) {
+            // ignore
+        }
+        expect(facade.chars).toEqual("A");
+    });
+});
+