+++ projects/w11k/tydux/src/lib/store-global.ts
@@ -2,13 +2,16 @@ import {TyduxStore} from "./store";
 
 let globalStore: TyduxStore | undefined;
 
-export function setGlobalStore(store: TyduxStore) {
-    if (globalStore !== undefined) {
+export function setGlobalStore(store: TyduxStore | undefined) {
+    if (globalStore !== undefined && store !== undefined) {
         throw new Error("a global store is already defined");
     }
     globalStore = store;
 }
 
-export function getGlobalStore(): TyduxStore | undefined {
+export function getGlobalStore(): TyduxStore {
+    if (globalStore === undefined) {
+        throw new Error("no global store registered");
+    }
     return globalStore;
 }