+++ projects/w11k/tydux/src/public_api.ts
@@ -4,7 +4,6 @@
 
 export * from "./lib/Facade";
 export { Commands } from "./lib/commands";
-export * from "./lib/view";
 export * from "./lib/development";
 export * from "./lib/store";
 export * from "./lib/store-global";